Version 1.2.0 beta - Annotation tracking
- Added annotation source tracking feature. Annotations are associated with a metadata object containing
information about their origin, whether it is auto-generated by certain method with certain parameters, or
added/modified by a human on certain date. The metatada is displayed in the status line (at bottom right
of the main screen) when hovering mouse over an annotation. In addition, there is a dialog (which is open via
menu View -> Annotaton Sources... or F6), which allows to gray out annotations associated with selected sources.
The dialog also has a "Comment" box to add user-defined comments on current edits.
The annotation source metadata is saved to regular *_contours.csv files as a specially formatted comment line,
this way maintaining compatibility with the previous versions, including the C++ (non-ML) one.